Add install target to Makefile
This will come into play later, as dpkg-buildpackage finds this, and
automatically inferrs that "pkgtut" should be shipped with the debian
package!

Hardcoding the path turns out to be a bad idea. When you run
"dpkg-buildpackage", it creates a working directory under the debian
directory. However, when debhelper calls "make install", it uses the
build system's path.

DESTDIR is a Makefile convention which allows for build systems to
"stage installs". You can read more about it here:

https://www.gnu.org/prep/standards/html_node/DESTDIR.html